handled lint error

+++ core/modules/rnn/encoder.py
@@ -35,13 +35,13 @@ class LSTMEncoder(nn.Module):
 
     x = encoded_inputs
     x = nn.Dropout(rate=self.dropout_rate)(x, deterministic=self.deterministic)
-    x = x.astype(cfg.dtype)
+    x = x.astype(self.dtype)
 
     # Input Encoder
     for layer_num in range(self.num_layers):
       initial_state = lstm.SimpleLSTM.initialize_carry((batch_size,), self.hidden_dim)
       _, x = lstm.SimpleLSTM(name=f"lstm_{layer_num}")(initial_state, x) # TODO(rgoel): Add layer norm to each layer
 
-    encoded = nn.LayerNorm(dtype=cfg.dtype, name='encoder_norm')(x)
+    encoded = nn.LayerNorm(dtype=self.dtype, name='encoder_norm')(x)
 
     return encoded
